Separate microbit-specific code out into separate modules.
New modules:
  display_control
  microbit_timer
  microbit_control
  microbit_matrix

Display is now generic over a Frame.

Frame is now a trait, with a MicrobitFrame implementation which client
code should work with instead of Frame.

Adds a new Matrix trait, describing the correspondence between the
visible layout of LEDs and the pins controlling them.

Adds a new DisplayControl trait (parallel to DisplayTimer), describing
how to manipulate the GPIO pins.

RowPlan is now public, so that Frame can be implemented.