Move non-microbit-specific code to a separate tiny-led-matrix crate.
Removes:
  display.rs
  display_control.rs
  display_timer.rs
  render.rs

Adds MicrobitGpio and MicrobitTimer1 newtypes for references to nrf51::GPIO
and nrf51::TIMER1, as the coherence rules will no longer let us implement the
DisplayControl and DisplayTimer traits directly on the nrf51 types.

Removes the blanket implementation of Render for T:Scrollable in favour of a
separate implementation on each Scrollable type (again forced by the coherence
rules).